backend/pyproject.toml
[project]
name = "auto-bidder-backend"
version = "0.1.0"
description = "Python AI service for auto-bidder platform"
authors = [{name = "Auto Bidder Team"}]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "datasets>=2.18",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "apscheduler>=3.10",
    "playwright>=1.41",
    "beautifulsoup4>=4.12",
    "httpx>=0.24",
]

[tool.uv]
dev-dependencies = ["pytest>=8", "pytest-asyncio>=0.23", "pytest-cov>=4", "ruff>=0.1", "black>=24", "mypy>=1.8"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]
asyncio_mode = "auto"
